+++ CHANGELOG.md
@@ -3,8 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 ### Fixed
-* Ensure that all group reports display correctly ([#487](https://github.com/YaleSTC/vesta/issues/487)).
 * Fix nav bar for special group members ([#474](https://github.com/YaleSTC/vesta/issues/474)).
+* Ensure that all group reports display correctly ([#487](https://github.com/YaleSTC/vesta/issues/487)).
+* Ensure that NextGroupsQuery ignores groups with no lottery number set ([#504](https://github.com/YaleSTC/vesta/issues/504)).
 
 ## v0.1.4 - 2017-03-29
 ### Fixed